drop payload_hash in favor of v5 payload_uuid
Turns out uuid v5 is what we were trying to get from payload_hash, and
allows us to use effectively the same hash in uuid form rather than
having to store an additional random uuid.

We can't perform a forward migration of existing payload_hash'es into v5
uuids, so the migration here truncates all table data (as actions relate
to payloads, everything must go). Typically this sort of change would
be a non-starter for a database migration, but given the early phase of
development for this project it seems like a reasonable action to take.